requirements.txt
Flask>=2.0
APScheduler>=3.9
python-dotenv>=1.0
yfinance>=0.2.25
requests>=2.28
python-telegram-bot>=13.15
nsetools>=1.0.11
pandas>=1.5
beautifulsoup4>=4.12